requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
openai>=1.50.0
httpx>=0.27.0
openpyxl>=3.1.5
xlrd>=2.0.1
xlwt>=1.3.0
python-docx>=1.1.2
olefile>=0.47
javalang>=0.13.0
loguru>=0.7.2
pydantic>=2.9.0
